// api/src/ai/tools/schema.ts
export const QueryInputSchema = z
    .object({
        fields: z.array(z.string()),
        sort: z.array(z.string()),
        filter: z.record(z.string(), z.any()),
        limit: z.number(),
        offset: z.number(),
        page: z.number(),
        search: z.string(),
        deep: z.record(z.string(), z.any()),
        alias: z.record(z.string(), z.string()),
        aggregate: z.object({
            count: z.array(z.string()).optional(),
            sum: z.array(z.string()).optional(),
            avg: z.array(z.string()).optional(),
            min: z.array(z.string()).optional(),
            max: z.array(z.string()).optional(),
        }),
        backlink: z.boolean(),
        version: z.string(),
        versionRaw: z.boolean(),
        export: z.string(),
        groupBy: z.array(z.string()),
    })
    .partial();

export const QueryValidateSchema = QueryInputSchema;
// ...
export const OperationItemInputSchema = z
    .object({
        id: z.string(),
        name: z.union([z.string(), z.null()]),
        key: z.string(),
        type: z.string(),
        position_x: z.number(),
        position_y: z.number(),
        options: z.record(z.string(), z.any()),
        resolve: z.union([z.string(), z.null()]),
        reject: z.union([z.string(), z.null()]),
        flow: z.string(),
        date_created: z.string(),
        user_created: z.string(),
    })
    .partial();

export const OperationItemValidateSchema = OperationItemInputSchema;

export const FlowItemInputSchema = z
    .object({
        id: z.string(),
        name: z.string(),
        icon: z.union([z.string(), z.null()]),
        color: z.union([z.string(), z.null()]),
        description: z.union([z.string(), z.null()]),
        status: z.enum(['active', 'inactive']),
        trigger: z.union([z.enum(['event', 'schedule', 'operation', 'webhook', 'manual']), z.null()]),
        options: z.union([z.record(z.string(), z.any()), z.null()]),
        operation: z.union([z.string(), z.null()]),
        operations: z.array(OperationItemInputSchema),
        date_created: z.string(),
        user_created: z.string(),
        accountability: z.union([z.enum(['all', 'activity']), z.null()]),
    })
    .partial();

export const FlowItemValidateSchema = FlowItemInputSchema;